tracy: move the multi-pass counter capture out of __main__.py
tools/tracy/perf_counter_multipass.py now owns the group tables, the pass scheduling, the architecture check, the per-pass replay and the device log merge. __main__.py keeps the CLI options and calls plan_perf_counter_capture and run_perf_counter_passes.
