+++ tests/ttnn/tracy/test_perf_counter_multipass.py
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from tracy.__main__ import (
+from tracy.perf_counter_multipass import (
     PERF_COUNTER_L1_GROUPS,
     PERF_COUNTER_MAX_GROUPS_PER_PASS,
     merge_perf_counter_device_logs,
